Rename folder to library
