+++ client/web/src/enterprise/batches/create/library/minimal.batch.yaml
